move 3d computations to superclass
